Split up contentItems selectors into separate files for easier lookup / testing.
