+++ app/modules/contentItems/index.js
@@ -3,7 +3,7 @@
 
 import * as actions from './actions';
 import * as model from './model';
-import * as selectors from './selectors';
+import selectors from './selectors';
 import components from './components';
 import reducer from './reducer';
 import saga from './saga';
