+++ app/modules/contentItems/saga/task/removeAndTogglePreviousItem.js
@@ -5,16 +5,16 @@ import ObjectNotFoundError from 'errors/usage-errors/ObjectNotFoundError';
 
 import * as t from '../../actionTypes';
 import { remove, toggleEditing } from '../../actions';
-import { getById, getAllById } from '../../selectors';
+import selectors from '../../selectors';
 import find from '../../lib/find';
 
 const removeAndTogglePreviousItemSaga = function* (
   action: t.RemoveAndTogglePreviousItemAction,
 ): Generator<*, *, *> {
-  const contentItemsById = yield select(getAllById);
+  const contentItemsById = yield select(selectors.getAllById);
   const { id } = action.payload;
 
-  const contentItemToRemove = yield select(getById, { id });
+  const contentItemToRemove = yield select(selectors.getById, { id });
   if (contentItemToRemove == null) throw new ObjectNotFoundError('contentItems:contentItem', id);
 
   // Find the previousEditorItem, if it exists. Do this before removing the contentItem.
