+++ app/modules/contentItems/saga/task/reverseIndent.js
@@ -6,16 +6,16 @@ import ObjectNotFoundError from 'errors/usage-errors/ObjectNotFoundError';
 import * as t from '../../actionTypes';
 import { move } from '../../actions';
 import { contentItemTypes } from '../../model';
-import { getById, getAllById } from '../../selectors';
+import selectors from '../../selectors';
 import find from '../../lib/find';
 
 const reverseIndentSaga = function* (action: t.ReverseIndentAction): Generator<*, *, *> {
   const { id } = action.payload;
 
-  const contentItemToReverseIndent = yield select(getById, { id });
+  const contentItemToReverseIndent = yield select(selectors.getById, { id });
   if (contentItemToReverseIndent == null) throw new ObjectNotFoundError('contentItems:contentItem', id);
 
-  const contentItemsById = yield select(getAllById);
+  const contentItemsById = yield select(selectors.getAllById);
   const parentOrSuperItem = find.parentOrSuperItem(contentItemToReverseIndent, contentItemsById);
 
   if (parentOrSuperItem != null) {
