+++ app/modules/contentItems/selectors/getAllById.js
@@ -1,0 +1,11 @@
+// @flow
+
+import type { State } from 'types/state';
+
+import type { ContentItemsById } from '../model';
+
+const getAllById = (state: State): ContentItemsById => {
+  return state.modules.contentItems.byId;
+};
+
+export default getAllById;
